+++ tests/Db/Schema/SchemaTest.php
@@ -6,9 +6,7 @@ namespace Yiisoft\Db\Tests\Db\Schema;
 
 use Yiisoft\Db\Cache\SchemaCache;
 use Yiisoft\Db\Constraint\Check;
-use Yiisoft\Db\Constraint\DefaultValue;
 use Yiisoft\Db\Constraint\ForeignKey;
-use Yiisoft\Db\Constraint\Index;
 use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Schema\Column\ColumnBuilder;
 use Yiisoft\Db\Schema\TableSchema;
@@ -42,38 +40,6 @@ final class SchemaTest extends IntegrationTestCase
         $this->assertSame([], Assert::invokeMethod($schema, 'findViewNames', ['dbo']));
     }
 
-    public function testGetSchemaChecks(): void
-    {
-        $db = $this->getSharedConnection();
-
-        $checks = [new Check('check_1', ['col1', 'col2'], 'col1 > col2')];
-        $schemaMock = $this->getMockBuilder(Schema::class)
-            ->onlyMethods(['findTableNames', 'loadTableChecks'])
-            ->setConstructorArgs([$db, TestHelper::createMemorySchemaCache()])
-            ->getMock();
-        $schemaMock->expects($this->once())->method('findTableNames')->willReturn(['T_constraints_1']);
-        $schemaMock->expects($this->once())->method('loadTableChecks')->willReturn($checks);
-        $tableChecks = $schemaMock->getSchemaChecks();
-
-        $this->assertSame(['T_constraints_1' => $checks], $tableChecks);
-    }
-
-    public function testGetSchemaDefaultValues(): void
-    {
-        $db = $this->getSharedConnection();
-
-        $defaultValues = [new DefaultValue('DF__T_constra__C_def__6203C3C6', ['C_default'], '((0))')];
-        $schemaMock = $this->getMockBuilder(Schema::class)
-            ->onlyMethods(['findTableNames', 'loadTableDefaultValues'])
-            ->setConstructorArgs([$db, TestHelper::createMemorySchemaCache()])
-            ->getMock();
-        $schemaMock->expects($this->once())->method('findTableNames')->willReturn(['T_constraints_1']);
-        $schemaMock->expects($this->once())->method('loadTableDefaultValues')->willReturn($defaultValues);
-        $tableDefaultValues = $schemaMock->getSchemaDefaultValues();
-
-        $this->assertSame(['T_constraints_1' => $defaultValues], $tableDefaultValues);
-    }
-
     public function testGetSchemaForeignKeys(): void
     {
         $db = $this->getSharedConnection();
@@ -97,22 +63,6 @@ final class SchemaTest extends IntegrationTestCase
         // The result is indexed by the name of the table the foreign keys belong to, see https://github.com/yiisoft/db/issues/1176
         $this->assertSame(['T_constraints_1'], array_keys($tableForeignKeys));
         $this->assertSame('T_constraints_2', $tableForeignKeys['T_constraints_1'][0]->foreignTableName);
-    }
-
-    public function testGetSchemaIndexes(): void
-    {
-        $db = $this->getSharedConnection();
-
-        $indexes = [new Index('PK__T_constr__A9FAE80AC2B18E65', ['"C_id'], true, true)];
-        $schemaMock = $this->getMockBuilder(Schema::class)
-            ->onlyMethods(['findTableNames', 'loadTableIndexes'])
-            ->setConstructorArgs([$db, TestHelper::createMemorySchemaCache()])
-            ->getMock();
-        $schemaMock->expects($this->once())->method('findTableNames')->willReturn(['T_constraints_1']);
-        $schemaMock->expects($this->once())->method('loadTableIndexes')->willReturn($indexes);
-        $tableIndexes = $schemaMock->getSchemaIndexes();
-
-        $this->assertSame(['T_constraints_1' => $indexes], $tableIndexes);
     }
 
     public function testGetSchemaNames(): void
@@ -144,44 +94,6 @@ final class SchemaTest extends IntegrationTestCase
         $this->assertTrue($schema->hasSchema('dbo'));
         $this->assertTrue($schema->hasSchema('public'));
         $this->assertFalse($schema->hasSchema('no_such_schema'));
-    }
-
-    public function testGetSchemaPrimaryKeys(): void
-    {
-        $db = $this->getSharedConnection();
-
-        $pksConstraint = new Index('PK__T_constr__A9FAE80AC2B18E65', ['"C_id'], true, true);
-        $schemaMock = $this->getMockBuilder(Schema::class)
-            ->onlyMethods(['findTableNames', 'getTablePrimaryKey'])
-            ->setConstructorArgs([$db, TestHelper::createMemorySchemaCache()])
-            ->getMock();
-        $schemaMock->expects($this->once())->method('findTableNames')->willReturn(['T_constraints_1']);
-        $schemaMock->expects($this->once())->method('getTablePrimaryKey')->willReturn($pksConstraint);
-        $tablePks = $schemaMock->getSchemaPrimaryKeys();
-
-        $this->assertIsArray($tablePks);
-        $this->assertContainsOnlyInstancesOf(Index::class, $tablePks);
-    }
-
-    public function testGetSchemaUniques(): void
-    {
-        $db = $this->getSharedConnection();
-
-        $uniquesConstraint = [new Index('CN_unique', ['C_unique'], true)];
-        $schemaMock = $this->getMockBuilder(Schema::class)
-            ->onlyMethods(['findTableNames', 'getTableUniques'])
-            ->setConstructorArgs([$db, TestHelper::createMemorySchemaCache()])
-            ->getMock();
-        $schemaMock->expects($this->once())->method('findTableNames')->willReturn(['T_constraints_1']);
-        $schemaMock->expects($this->once())->method('getTableUniques')->willReturn($uniquesConstraint);
-        $tableUniques = $schemaMock->getSchemaUniques();
-
-        $this->assertIsArray($tableUniques);
-
-        foreach ($tableUniques as $uniques) {
-            $this->assertIsArray($uniques);
-            $this->assertContainsOnlyInstancesOf(Index::class, $uniques);
-        }
     }
 
     public function getTableSchema(): void